requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beanie==1.23.6
motor==3.3.2
anthropic==0.7.8
openai==1.3.8
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0